Move common functions to tion class
