+++ tion/__init__.py
@@ -1,2 +1,3 @@
-from .tion import tion
+from .tion import tion, TionException
 from .s3 import s3
+
